update on vec-set
